+++ src/makielayout/blocks.jl
@@ -615,3 +615,52 @@ function REPL.fielddoc(t::Type{<:Block}, s::Symbol)
     default_str = Makie.attribute_default_expressions(t)[s]
     return repl_docstring(nameof(t), s, docs, examples, default_str)
 end
+
+"""
+    function tooltip!(b::Block, str::AbstractString; placement=:above, kwargs...)
+
+Adds a tooltip to a block.  See `tooltip` for more details.  To make it only
+appear when hovering, set the `visible` property of the tooltip to the
+`hovering` property of the block.
+
+# Examples
+```julia-repl
+julia> f = Figure()
+
+julia> t = Toggle(f[1,1])
+Toggle()
+
+julia> tooltip!(t, "I'm a Toggle", visible = t.hovering)
+Plot{Makie.tooltip, Tuple{Vec{2, Float32}, String}}
+
+julia> b = Button(f[2,1])
+Button()
+
+julia> tt = tooltip!(b, "I'm a Button", placement = :below)
+Plot{Makie.tooltip, Tuple{Vec{2, Float32}, String}}
+
+julia> on(h -> tt.visible[]=h, b.hovering)
+ObserverFunction defined at REPL[7]:1 operating on Observable(false)
+```
+"""
+function tooltip!(b::Block, str::AbstractString; placement=:above, kwargs...)
+    position = lift(b.layoutobservables.computedbbox) do bbox
+        if placement == :above
+            bbox.origin + Point2f((bbox.widths[1]/2, bbox.widths[2]))
+        elseif placement == :below
+            bbox.origin + Point2f((bbox.widths[1]/2, 0))
+        elseif placement == :left
+            bbox.origin + Point2f((0, bbox.widths[2]/2))
+        elseif placement == :right
+            bbox.origin + Point2f((bbox.widths[1], bbox.widths[2]/2))
+        elseif placement == :center
+            bbox.origin + Point2f((bbox.widths[1]/2, bbox.widths[2]/2))
+        else
+            @error("invalid value for tooltip_placement, using :above")
+            bbox.origin + Point2f((bbox.widths[1]/2, bbox.widths[2]))
+        end
+    end
+    tt = tooltip!(b.blockscene, position, str; placement, kwargs...)
+    translate!(tt, 0, 0, 9e3)
+    return tt
+end