+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Add Tooltips for blocks [#4451](https://github.com/MakieOrg/Makie.jl/pull/4451)
 - For `Textbox`es in which a fixed width is specified, the text is now scrolled
   if the width is exceeded [#4293](https://github.com/MakieOrg/Makie.jl/pull/4293)
 - Changed image, heatmap and surface picking indices to correctly index the relevant matrix arguments. [#4459](https://github.com/MakieOrg/Makie.jl/pull/4459)